validation formulaire ok

+++ front-end/js/panier.js
@@ -1,48 +1,43 @@
-import { initStorage } from "./storage.js";
-import { cleanStorage } from "./storage.js";
+import { Cart } from "./cart.js";
 import { sendOrder } from "./api.js";
-import { validateName } from "./regex.js";
-import { validateMail } from "./regex.js";
-import { validateAddress } from "./regex.js";
+import { testForm } from "./regex.js";
+//import { validateMail } from "./regex.js";
+//import { validateAddress } from "./regex.js";
+import { Camera } from "./camera.js";
 
 function displayStorage() {
-  //si le local storage est vide on va afficher un texte a l'écran et cacher le formulaire
   if (localStorage.length === 0) {
-    let p = document.createElement("p");
-    p.textContent = "Il n'y a actuellement aucun article dans votre panier.";
-    document.getElementById("panier").appendChild(p);
-    document.getElementById("formulaire").style.display = "none";
+    //si le localStorage est vide:
+    emptyCartDisplay();
   }
   //sinon on va afficher le panier ainsi qu'un bouton pour le vider et la somme du panier
   else {
-    let storage = JSON.parse(window.localStorage.getItem("camera"));   
-    for (let i = 0; i < storage.length; i++) {
-      let camera = storage[i];
-      displayCart(camera, storage);  
+    let cart = new Cart();
+    let cartArray = cart.getItems();
+    for (let i = 0; i < cartArray.length; i++) {
+      let camera = cartArray[i];
+      displayCart(camera, cartArray);
     }
     cleanCart();
-    createTotalPrice(storage);
+    createTotalPrice();
   }
 }
 
-//-----fonction qui va afficher le prix total du panier------------a faire
-function cartFinalSum(array) {
-  let priceArray = [];
-  array.forEach(camera => getPrice(camera, priceArray));
-  let cartSum = priceArray.reduce((a, b) => a + b);
-  document.querySelector("#cartTools p").textContent = "Montant du panier : " + cartSum + " €";
+//-----fonction qui va afficher un texte a l'écran et cacher le formulaire--------
+function emptyCartDisplay() {
+  let p = document.createElement("p");
+  p.textContent = "Il n'y a actuellement aucun article dans votre panier.";
+  document.getElementById("panier").appendChild(p);
+  document.getElementById("formulaire").style.display = "none";
 }
 
-function getPrice(camera, array){
-  let totalPrice = (camera.price / 100) * camera.cameraCount;
-  array.push(totalPrice);
-  return array;
-}
-
-function createTotalPrice(array){
+//-----fonction qui va afficher le prix total du panier---------------------
+function createTotalPrice() {
   let sum = document.createElement("p");
   document.getElementById("cartTools").appendChild(sum);
-  cartFinalSum(array);
+  let totalSum = Cart.getTotalValue();
+  console.log(totalSum);
+  sum.textContent = "Montant du panier : " + totalSum +" €";
 }
 
 //---------------------fonction qui va enler un article de la quantité----------------
@@ -51,7 +46,7 @@ function removeItem(camera, array) {
     let index = array.indexOf(camera);
     if (index > -1 && array.length > 1) {
       array.splice(index, 1);
-      initStorage(array);
+      Cart.initCart(array);
       window.location.reload();
     } else if (index > -1 && array.length == 1) {
       window.localStorage.clear();
@@ -59,7 +54,7 @@ function removeItem(camera, array) {
     }
   } else {
     camera.cameraCount--;
-    initStorage(array);
+    Cart.initCart(array);
     updateCardBottom(camera, array);
   }
 }
@@ -67,21 +62,24 @@ function removeItem(camera, array) {
 //--------------------fonction qui va ajouter un article de la quantité-------------------
 function addItem(camera, array) {
   camera.cameraCount++;
-  initStorage(array);
-  updateCardBottom(camera, array); 
+  Cart.initCart(array);
+  updateCardBottom(camera, array);
 }
 
 //----fonction qui va modifier la quantité et le prix sur la carte affiché----------------
-function updateCardBottom(camera, array){
+function updateCardBottom(camera, array) {
   let quantity = document.getElementsByClassName("quantity");
   let price = document.getElementsByClassName("totalPricePerItem");
   let index = array.indexOf(camera);
   quantity[index].textContent = "quantité : " + camera.cameraCount;
-  price[index].textContent = "Prix total : " + (camera.price / 100) * camera.cameraCount + " €";
-  cartFinalSum(array);
+  price[index].textContent =
+    "Prix total : " +
+    Cart.formatedPrice(camera.price) * camera.cameraCount +
+    " €";
+  //cartFinalSum(array);
 }
 
-//-----fonction qui crée une carte pour chaque camera du panier-------------------
+//------------fonction qui crée une carte pour chaque camera du panier-------------------
 function createCard() {
   const panier = document.getElementById("panier");
   let card = document.createElement("div");
@@ -101,7 +99,7 @@ function createCardHeader(camera) {
   title.innerHTML = camera.name;
   cardHeader.appendChild(title);
   let price = document.createElement("p");
-  price.textContent = camera.price / 100 + " €";
+  price.textContent = Cart.formatedPrice(camera.price) + " €";
   cardHeader.appendChild(price);
   return cardHeader;
 }
@@ -131,16 +129,18 @@ function createCardBottom(camera, array) {
   removeButton.textContent = "-";
   removeButton.classList.add("buttn");
   removeButton.classList.add("removeBtn");
-  quantityDiv.appendChild(removeButton);  
+  quantityDiv.appendChild(removeButton);
   let addButton = document.createElement("button");
   addButton.textContent = "+";
   addButton.classList.add("buttn");
   addButton.classList.add("addBtn");
-  quantityDiv.appendChild(addButton);  
+  quantityDiv.appendChild(addButton);
   let totalPricePerItem = document.createElement("p");
   totalPricePerItem.classList.add("totalPricePerItem");
   totalPricePerItem.textContent =
-    "Prix total : " + (camera.price / 100) * camera.cameraCount + " €";
+    "Prix total : " +
+    Cart.formatedPrice(camera.price) * camera.cameraCount +
+    " €";
   cardContent.appendChild(totalPricePerItem);
   removeButton.addEventListener("click", () => removeItem(camera, array));
   addButton.addEventListener("click", () => addItem(camera, array));
@@ -161,13 +161,12 @@ function displayCart(camera, array) {
 //-----------------fonction qui crée un bouton pour vider le panier------------------------
 function cleanCart() {
   let div = document.createElement("div");
-  let tool = document.getElementById("cartTools");
-  tool.appendChild(div);
+  document.getElementById("cartTools").appendChild(div);
   let cleanButton = document.createElement("button");
   cleanButton.classList.add("bg-primary");
   cleanButton.textContent = "Vider le panier";
   div.appendChild(cleanButton);
-  cleanButton.addEventListener("click", () => cleanStorage());
+  cleanButton.addEventListener("click", () => Cart.cleanCart());
 }
 
 //fonction qui apres avoir les données de l'api va vider le panier et envoyer l'utilisateur sur
@@ -175,60 +174,48 @@ function cleanCart() {
 async function receiveOrder(data) {
   const orderReturn = await sendOrder(data);
   window.location.href =
-    "validation.html?name=" + orderReturn.contact.firstName + "&orderId=" + orderReturn.orderId;
+    "validation.html?name=" +
+    orderReturn.contact.firstName +
+    "&orderId=" +
+    orderReturn.orderId;
   window.localStorage.clear();
 }
 
-let sendButton = document.getElementById("sendButton");
-sendButton.addEventListener("click", function (e) {
-  e.preventDefault();
+//----------fonction qui va créer l'objet contact---------------------
+function getFormData() {
   let lastName = document.getElementById("lastName").value;
   let firstName = document.getElementById("firstName").value;
   let email = document.getElementById("email").value;
   let address = document.getElementById("address").value;
   let city = document.getElementById("city").value;
   let contact = { lastName, firstName, email, address, city };
-  let storage = JSON.parse(window.localStorage.getItem("camera"));
-  let products = [];
-  storage.forEach((camera) => {
-    let cameraIdToSend = camera._id;
-    products.push(cameraIdToSend);
+  return contact;
+}
+
+//--fonction qui va tester les données du formulaire puis si elles sont valides--------
+//--------va créer les objets a envoyer puis envoyer les données a l'api
+function activateSendButton() {
+  let button = document.getElementById("sendButton");
+  button.addEventListener("click", (e) => {
+    e.preventDefault();
+    let error = testForm();
+    if (error.length !== 0) {
+      alert(error);
+    } else {
+      let contact = getFormData();
+      let products = [];
+      let cart = new Cart;
+      let storage = cart.getItems();
+      storage.forEach((camera) => {
+        let cameraIdToSend = camera._id;
+        products.push(cameraIdToSend);
+      });
+      let order = { contact, products };
+      receiveOrder(order);
+      window.localStorage.clear();
+    }
   });
-  let order = { contact, products };
-  receiveOrder(order);
-  window.localStorage.clear();
-});
+}
 
-//-------------------------------regex-----------------------------------
-const fieldLastName = document.getElementById("lastName");
-fieldLastName.addEventListener("change", (e) => validateName(e.target.value, fieldLastName));
-
-let fieldfirstName = document.getElementById("firstName");
-fieldfirstName.addEventListener("change", (e) => validateName(e.target.value, fieldfirstName));
-
-let fieldEmail = document.getElementById("email");
-fieldEmail.addEventListener("change", (e) => validateMail(e.target.value, fieldEmail));
-
-let fieldAddress = document.getElementById("address");
-fieldAddress.addEventListener("change", (e) => validateAddress(e.target.value, fieldAddress));
-
-let fieldCity = document.getElementById("city");
-fieldCity.addEventListener("change", (e) => validateName(e.target.value, fieldCity));
-
-let input = document.querySelector("input");
-let button = document.getElementById("sendButton");
-//button.disabled = true;
-input.addEventListener("change", (e) => {
-  console.log("test");
-  if (validateName(e.target.value, fieldLastName) == true && 
-      validateName(e.target.value, fieldLastName) == true &&
-      validateMail(e.target.value, fieldEmail) == true &&
-      validateAddress(e.target.value, fieldAddress) == true &&
-      validateName(e.target.value, fieldCity) == true){
-        console.log("regex ok")
-        button.disabled = false;
-      }
-});
-//----------------------fin regex-----------------------------------
-
+activateSendButton();
 displayStorage();
